feat: All apps are running in one cluster. Ingress is working and skaffold too.
I created basic k8s config files. MOst of them are depl file where i have deployment(type of object in k8s which makes sure that everything is running in right number and succesfull) and IPcluster (object which makes sure that pod(container basicly but in k8s infra) is exposed to local infra of cluster) and for posts there is one config file for NodePort(which makes sure that it is available to world. There is ingress for makeing routes to pods inside cluster and makeing ngninx server on posts.com edited in /etc/zshrc file. There is also skaffold file at the root which is great tool for not doing docker build, push; kubectl(k) apply, rollout; but instead inside root / dir, type "skaffold dev" and this is automatically going to run all pods and what over changes and also print everything in one shell. if i make change, skaffold is going to see it and if it is under the manual-src path then it is throwing this file to pod and if it is not then it is rebuilding and pushing image and rebuild deployment and everything what is related to project